Type Shorthands

The package defines and exports shorthands for commonly used types. They are defined as

I64 = Int64
F64 = Float64
CF64 = ComplexF64
VecF64 = Vector{Float64}
VecCF64 = Vector{ComplexF64}
MatF64 = Matrix{Float64}
MatCF64 = Matrix{ComplexF64}

Warning

This is bad practice, but often more convenient when quickly prototyping...

cSpline

Interpolation of complex-valued data using Dierckx.jl. We provide both a functor, and a classical interface.

PConv.cSpline — Type

cSpline(x::AbstractVector{T}, y::AbstractVector; w=ones(length(x)), k=3, bc="nearest", s=0.0)

Struct and Functor for interpolating complex data, using Dierckx.jl's Spine1D. Spline can be evaluated by calling the struct as a function. Regular derivative and integrate functions are avaivible.

Arguments

k degree of spline k (1 = linear, 2 = quadratic, 3 = cubic, up to 5)

bc{"nearest" (default), "zero", "extrapolate", "error"} Behavior when evaluating

the spline outside the support domain (minimum(x), maximum(x)).

PConv.cspline — Method

cspline(xi, x::AbstractArray, y::AbstractArray; kwargs...)

Complex Interpolation function, by calleing Spline1D for Re and Im seperately. Returns Interpolation of x,y evaluated at xi, if y is a complex array.

Arguments

xi::AbstractArray: New points

x::AbstractArray{Real}: x-values of the data (ordered and real)

y::AbstractArray: y-values of the data ∈ Complex

kwargs::NamedTuple: Stanard options for Spline1D

return: Spline1D(x, y, bc)(xi)

PConv.dderivative_cspline — Method

dderivative_cspline(xy_eval::Tuple, x, A::AbstractMatrix; k = 3, nu_x = 1, nu_y = 1)

Calculate derivate of complex matrix A(x, x) at xy_eval using Dierckx.

PConv.derivative_cspline — Method

derivative_cspline(x_eval, x, fx; k = 3, nu = 1)

Calculate derivate of complex function fx at x_eval using Dierckx.

Physics

Physics related Helper functions

PConv.heaviside — Function

heaviside(x::Real, zero::Real=1.0)

Heaviside function Θ(x) with optional zero-point argument θ(0.0)=zero.

Stanard value (zero=1.0) implements Θ(x+0⁺).

PConv.nb — Method

nb(x::T1, β::T2=1.0) where {T1 <: Number, T2 <: Number}

Bose function $nf(x) = 1 / (e^(x*β) - 1)$ at x with temperature β.

PConv.nf — Method

nf(x::T1, β::T2=1.0) where {T1 <: Number, T2 <: Number}

Fermi function $nf(x) = 1 / (e^(x*β) + 1)$ at x with temperature β.

PConv.pauliMat — Method

pauliMat(x::Int)

Pauli matrices with x=0,...,3, defined by

'x=0: [1  0; 0 1] ; x=1 [0 1; 1 0 ]'\
'x=2: [0 -i; i 0] ; x=3 [1 0; 0 -1]'

return: σˣ

Convenience

More general convenience functions, that are often useful

Base.identity — Method

identity(x::Vararg{Any, N}) where N

Extend Base.identity function to accept multiple arguments

return: x

PConv.:≂ — Method

@generated function ≂(x, y)

Lenient comparison operator for struct, both mutable and immutable (type with \eqsim).

PConv.fdiff — Method

fdiff(a, b) = findmax(@. abs(a - b))

Convenience function for calculating the index and value of maximal difference.

PConv.generate_interval — Method

generate_interval(t, t_start, t_step, t_end; digits = 8)
generate_interval(t, t_values; digits = 8)

Generate index-array for t_values in t. If the exact value is not contained take the closest value that is.

PConv.generate_sym — Method

generate_sym(Xmin, Xmax, δx)
generate_sym(Xmax, δx)

Generate symmetric array around 0, i.e. X = Xmin, ..., δx, 0, δx, ..., Xmax. If Xmin is not provided, Xmin = -Xmax is used. Note that X[1] != Xmin , X[end] != Xmax is possible if δx does not fit into the interval.

Arguments

Xmin(=-Xmax): Minimal value

Xmax: Maximal value

δx: Spacing

return: X, N, N0 (X[N0] = 0.0)

PConv.isapprox_eps — Method

isapprox_eps(x::Complex{T}, y::Complex{T}; factor::Integer=1) where {T<:Real}
isapprox_eps(x::T, y::T; factor::Integer=1) where {T<:Real}

Compare x, y with times⋅eps(max(x, y)) accuracy, where eps(x) gives floating point accuracy.

In the case of complex numbers compare real and imag part seperately.

Returns

Bool Boolian if x ≈ y with accuracy times⋅eps(max(x, y))

PConv.match_lengths — Method

match_lengths(args...)

Matches the lengths of the input arrays by padding singleton arrays (numbers) to matching length.

Arguments

args...: Variable number of arrays to be matched in length. (length of array matches or is 1)

Returns

An array of arrays with lengths adjusted to match the maximum length among the input arrays.

PConv.mdiff — Method

mdiff(a, b) = maximum(@. abs(a - b))

Convenience function for calculating the maximal difference.

PConv.minimize_distance — Method

minimize_distance(x::Abstractarray, y::Abstractarray)

Determine permutation indices such that ||x⃗ - y⃗|| ∀i is minimal Note: using sum(abs2, .) instead of norm in order to not introduce an additional dependence!

PConv.search_nearest — Method

search_nearest(a::AbstractArray, x; sorted::Bool = false)

Find index of closest element ∈ a to x. sorted decides if a is sorted.

PConv.search_nearest_sorted — Method

search_nearest_sorted(a::AbstractArray, x)

Find index of closest element ∈ a to x. a is assumed sorted.

PConv.size_mb — Method

size_mb(x)

Print size of x in mb.

PConv.void — Method

void(x::Vararg{Any, N}) where N

Take arbritrary number of arguments and return nothing.

IO

Convenience functions for generating and handling (versioned) paths

PConv.PROJECT_ROOT — Constant

Absolute path of current active project (folder containing Project.toml) at the time of including using PConv

PConv.dirnames — Method

dirnames(n::Integer, path::AbstractString)

Extend Base.dirname function to direcly apply it n-times

PConv.find_project_root — Method

find_project_root(folder_path::AbstractString)

Find the root-path of project by searching for first Project.toml above/equal folder_path.

PConv.overprint — Method

overprint(str::AbstractString)

Replace current line with str, i.e. overprint current line.

PConv.versionized_path — Method

versionized_path(filename; abspath=PROJECT_ROOT, overwrite = false)

Generate versionized path*filename by appending _vx to the filename if filename already exists.

Arguments

filename::AbstractString: Folder+name of the file with file extension (Example "pictures/test.png")

abspath::AbstractString=PROJECT_ROOT: Absolute path to use. Default is the

directory of the active project.

overwrite::Bool If false, don't versionize the path and potentially overwrite existing

file

return: versionized_path::AbstractString